requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0

# Authentication & Security
bcrypt==4.0.1
Werkzeug==2.3.7
WTForms==3.0.1

# API & Serialization
Flask-RESTful==0.3.10
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# PDF Generation
reportlab==4.0.4
playwright==1.55.0

# Background Tasks
celery==5.3.1
redis==4.6.0

# Production Server
gunicorn==21.2.0

# Cloud Storage
boto3==1.34.131
google-cloud-storage==2.10.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
python-dateutil==2.8.2

# Development
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
